added DOM Element functionality for edit's selector parameter
+ edit now accepts a DOM Element for the "selector" attribute (as well
as being able to pass a selector like before
+ you may now omit the "inputType" parameter and specify "parent" in its
place

Did this for one reason: to add point-and-click editing functionality.
By attaching a "click" event listener to the document and calling:
scribly.edit(e.target, [parent]);
as its callback you can allow your users to edit inline elements on the
fly with a single click.